move url and public settings back to general
