+++ app/(main)/o/[slug]/settings/general-settings.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Loader2 } from "lucide-react";
+import { Check, Copy, Loader2 } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { useMemo, useState, useEffect } from "react";
 import { useForm, UseFormReturn } from "react-hook-form";
@@ -13,8 +13,10 @@ import { AutosizeTextarea } from "@/components/ui/autosize-textarea";
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
+import { Switch } from "@/components/ui/switch";
 import { updateTenantSchema } from "@/lib/api";
 import { DEFAULT_WELCOME_MESSAGE } from "@/lib/constants";
+import { getSettingsPath } from "@/lib/paths";
 import * as schema from "@/lib/server/db/schema";
 import { cn } from "@/lib/utils";
 
@@ -23,12 +25,27 @@ import { HelpWelcomeMessageDialog } from "./help-welcome-message-dialog";
 // Transform null to empty string for form field handling
 const nullToEmptyString = (v: string | null) => v ?? "";
 
+const isValidSlug = (slug: string) => {
+  if (!slug || slug.trim().length === 0) return false;
+  const sanitized = slug
+    .trim()
+    .toLowerCase()
+    .replace(/[^\w\s-]/g, "")
+    .replace(/[\s_-]+/g, "-")
+    .replace(/^-+|-+$/g, "");
+  return slug === sanitized;
+};
+
 const formSchema = z.object({
   question1: z.string().nullable().transform(nullToEmptyString),
   question2: z.string().nullable().transform(nullToEmptyString),
   question3: z.string().nullable().transform(nullToEmptyString),
   welcomeMessage: z.string().nullable().default(DEFAULT_WELCOME_MESSAGE).transform(nullToEmptyString),
   name: z.string().min(1, "Name must be at least 1 character").max(30, "Name must be less than 30 characters"),
+  slug: z.string().nullable().transform(nullToEmptyString).refine(isValidSlug, {
+    message: "URL can only contain lowercase letters, numbers, and hyphens",
+  }),
+  isPublic: z.boolean().default(false),
 });
 
 type FormValues = z.infer<typeof formSchema>;
@@ -52,6 +69,7 @@ const QuestionField = ({ form, name, label }: QuestionFieldProps) => (
             placeholder="Type something"
             className="rounded-[8px] border border-[#D7D7D7] h-[58px] placeholder-[#74747A] text-[16px]"
             {...field}
+            value={String(field.value)}
           />
         </FormControl>
         <FormMessage />
@@ -109,7 +127,7 @@ const TextAreaField = ({ form, name, label, className, help, hasDefault }: TextA
           </FormLabel>
           <FormControl>
             <div className="rounded-md border border-input bg-transparent px-3 py-1 text-base shadow-sm transition-colors placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:cursor-not-allowed disabled:opacity-50 md:text-sm">
-              <AutosizeTextarea className="pt-1.5" minHeight={80} {...field} />
+              <AutosizeTextarea className="pt-1.5" minHeight={80} {...field} value={String(field.value)} />
             </div>
           </FormControl>
           <FormMessage />
@@ -138,6 +156,7 @@ const CompanyNameField = ({ form, name, label }: CompanyNameFieldProps) => {
               type="text"
               className="rounded-[8px] border border-[#D7D7D7] h-[58px] placeholder-[#74747A] text-[16px]"
               {...field}
+              value={String(field.value)}
             />
           </FormControl>
           <FormMessage />
@@ -146,6 +165,93 @@ const CompanyNameField = ({ form, name, label }: CompanyNameFieldProps) => {
     />
   );
 };
+
+type URLFieldProps = {
+  name: keyof FormValues;
+  label: string;
+  form: UseFormReturn<FormValues>;
+};
+
+const URLField = ({ form, name, label }: URLFieldProps) => {
+  const [isCopied, setIsCopied] = useState(false);
+  const [origin, setOrigin] = useState("");
+
+  useEffect(() => {
+    setOrigin(window.location.origin);
+  }, []);
+
+  const handleCopyUrl = () => {
+    const url = `${origin}/o/${form.getValues(name) || "your-chat-name"}`;
+    navigator.clipboard.writeText(url);
+    setIsCopied(true);
+    setTimeout(() => setIsCopied(false), 2000);
+  };
+
+  return (
+    <FormField
+      control={form.control}
+      name={name}
+      render={({ field }) => (
+        <FormItem className="flex flex-col mt-8">
+          <FormLabel className="font-semibold text-[16px] mb-3">{label}</FormLabel>
+          <FormControl>
+            <Input
+              type="text"
+              placeholder="Enter URL"
+              className="rounded-[8px] border border-[#D7D7D7] h-[58px] placeholder-[#74747A] text-[16px]"
+              {...field}
+              value={String(field.value)}
+            />
+          </FormControl>
+          <div className="text-sm text-muted-foreground mt-1 flex items-center gap-2">
+            <span>
+              Your chat will be available at: {origin}/o/{field.value || "your-chat-name"}
+            </span>
+            <button
+              type="button"
+              onClick={handleCopyUrl}
+              className="p-1 hover:bg-gray-100 rounded-md transition-colors"
+              title="Copy URL"
+            >
+              {isCopied ? <Check className="h-4 w-4 text-green-500" /> : <Copy className="h-4 w-4 text-black" />}
+            </button>
+          </div>
+          <FormMessage />
+        </FormItem>
+      )}
+    />
+  );
+};
+
+type SwitchFieldProps = {
+  name: keyof FormValues;
+  label: string;
+  form: UseFormReturn<FormValues>;
+};
+
+const SwitchField = ({ form, name, label }: SwitchFieldProps) => (
+  <FormField
+    control={form.control}
+    name={name}
+    render={({ field }) => (
+      <FormItem className="flex flex-row items-center justify-between mt-8">
+        <div className="space-y-0.5">
+          <FormLabel className="font-semibold text-[16px]">{label}</FormLabel>
+          <p className="text-sm text-muted-foreground">Anyone with the link can chat with your AI assistant</p>
+        </div>
+        <div className="flex-shrink-0 ml-4">
+          <FormControl>
+            <Switch
+              checked={Boolean(field.value)}
+              onCheckedChange={field.onChange}
+              className="data-[state=checked]:bg-[#D946EF]"
+            />
+          </FormControl>
+        </div>
+      </FormItem>
+    )}
+  />
+);
 
 type Props = {
   tenant: typeof schema.tenants.$inferSelect;
@@ -181,6 +287,29 @@ export default function GeneralSettings({ tenant, canUploadLogo }: Props) {
     setLoading(true);
 
     try {
+      // If the slug has changed, check if it's available
+      if (values.slug !== tenant.slug) {
+        const checkResponse = await fetch("/api/tenants/check-slug", {
+          method: "POST",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({
+            slug: values.slug,
+            tenantId: tenant.id,
+          }),
+        });
+
+        if (!checkResponse.ok) {
+          throw new Error("Failed to check slug availability");
+        }
+
+        const { available } = await checkResponse.json();
+        if (!available) {
+          toast.error("This URL is already taken. Please choose a different one.");
+          setLoading(false);
+          return;
+        }
+      }
+
       const payload = updateTenantSchema.parse(values);
       const res = await fetch("/api/tenants/current", {
         method: "PATCH",
@@ -197,6 +326,11 @@ export default function GeneralSettings({ tenant, canUploadLogo }: Props) {
         ...values,
         welcomeMessage: values.welcomeMessage.length ? values.welcomeMessage : undefined,
       });
+
+      // If the slug was changed, redirect to the new URL
+      if (values.slug !== tenant.slug) {
+        router.push(getSettingsPath(values.slug));
+      }
     } catch (error) {
       toast.error("Failed to save changes");
       console.error(error);
@@ -268,6 +402,11 @@ export default function GeneralSettings({ tenant, canUploadLogo }: Props) {
             <QuestionField form={form} name="question2" label="Question 2" />
             <QuestionField form={form} name="question3" label="Question 3" />
 
+            <hr className="w-full my-8" />
+
+            <URLField form={form} name="slug" label="URL name" />
+            <hr className="w-full my-8" />
+            <SwitchField form={form} name="isPublic" label="Enable public chat" />
             <hr className="w-full my-8" />
           </div>
         </form>
